refactor: ViewModel에서 Model에 해당되는 코드 분리
